requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
requests==2.32.3
PyJWT==2.8.0